% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyHelpers.R
\name{build_key}
\alias{build_key}
\title{Build CatMapper Key Expressions}
\usage{
build_key(field, value)
}
\arguments{
\item{field}{Field name(s) used on the left side of the expression.}

\item{value}{Value(s) used on the right side of the expression.}
}
\value{
A character vector of key expressions.
}
\description{
Build key expressions in the \code{FIELD == VALUE} format used by CatMapper
upload workflows.
}
\examples{
build_key("Type", "Adamana Brown")
build_key(c("Type", "Region"), c("Adamana Brown", "Flagstaff"))
}
