% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{callAPI}
\alias{callAPI}
\title{Call API}
\usage{
callAPI(
  endpoint,
  parameters,
  request = "GET",
  url = NULL,
  type = "default",
  headers = NULL
)
}
\arguments{
\item{endpoint}{API endpoint}

\item{parameters}{API parameters}

\item{request}{GET or POST}

\item{url}{API URL override. If \code{NULL}, \code{CATMAPR_API_URL} is used when set, otherwise \code{"https://api.catmapper.org"}.}

\item{type}{default or stream}

\item{headers}{Optional named list of request headers, for example \code{list("X-API-Key" = "cmk_...")}.}
}
\value{
API response
}
\description{
Helper function to call the CatMapper API.
}
\examples{
\dontrun{
CatMapR:::callAPI(
  endpoint = "search",
  parameters = list(
    term = "Dan",
    database = "SocioMap",
    property = "Name",
    domain = "ETHNICITY"
  ),
  request = "GET"
)
}

}
\keyword{internal}
