% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMID.R
\name{CMIDinfo}
\alias{CMIDinfo}
\title{Retrieve Details for a Specific CMID (CatMapperID)}
\usage{
CMIDinfo(database, cmid)
}
\arguments{
\item{database}{A string specifying the database to search in. Valid options are "SocioMap" or "ArchaMap".}

\item{cmid}{The CatMapperID for which details are to be retrieved (e.g., "SM1" or "AM1").}
}
\value{
A list containing node properties and relationships associated with the specified CMID, or an error message if the request fails.
}
\description{
This function retrieves detailed information about a specified CatMapperID (CMID) from a specified database.
It fetches both node properties and their relationships associated with the given CMID.
}
\examples{
CMIDinfo(database = "SocioMap", cmid = "SM1")
CMIDinfo(database = "ArchaMap", cmid = "AM1")
}
