% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{getUploadProperties}
\alias{getUploadProperties}
\title{Retrieve Upload Property Metadata from a Specified Database}
\usage{
getUploadProperties(database = "SocioMap", url = NULL)
}
\arguments{
\item{database}{A string specifying the database from which to retrieve upload property metadata. Valid options are \code{"SocioMap"} or \code{"ArchaMap"}.}

\item{url}{API URL override. If \code{NULL}, \code{CATMAPR_API_URL} is used when set, otherwise \code{"https://api.catmapper.org"}.}
}
\value{
A list with three elements:
\describe{
\item{database}{Database name returned by the API.}
\item{nodeProperties}{A data frame of node property metadata.}
\item{usesProperties}{A data frame of USES relationship property metadata.}
}
}
\description{
This function retrieves upload-oriented CatMapper property metadata for a
specified database. The API groups results into node properties and USES
relationship properties.
}
\examples{
\dontrun{
getUploadProperties(database = "SocioMap")
getUploadProperties(database = "ArchaMap")
}
}
