% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{searchDatabase}
\alias{searchDatabase}
\title{Search}
\usage{
searchDatabase(
  database,
  domain = NULL,
  term = NULL,
  property = "Name",
  yearStart = NULL,
  yearEnd = NULL,
  country = NULL,
  context = NULL,
  dataset = NULL,
  query = "false",
  limit = 1000
)
}
\arguments{
\item{database}{name of database (SocioMap or ArchaMap)}

\item{domain}{name of category domain to search for (DISTRICT, ETHNICITY, etc.)}

\item{term}{search term ("Afghanistan")}

\item{property}{property to search by (Name, CMID, or Key)}

\item{yearStart}{year to search by start}

\item{yearEnd}{year to search by end}

\item{country}{country to search by (must be a CMID)}

\item{context}{context to search by (e.g., hierarchical category as in the state for a county--must be a CMID)}

\item{dataset}{dataset to search by (must be a dataset CMID)}

\item{query}{return query instead of results ('true' or 'false')}

\item{limit}{limit number of results (currently ignored by API and capped at 10,000 server-side)}
}
\value{
A list with \code{data} and \code{count} elements.
}
\description{
Search for a term in a database and retrieve the CMID (CatMapperID), CMName (CatMapper Name), and other information about the matches.
}
\examples{
\dontrun{
searchDatabase(database = "SocioMap", domain = "ETHNICITY", term = "Dan", property = "Name")
}
}
