% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allDatasets.R
\name{allDatasets}
\alias{allDatasets}
\title{Retrieve All Datasets from a Specified Database}
\usage{
allDatasets(database)
}
\arguments{
\item{database}{A string specifying the database from which to retrieve datasets. Valid options are "SocioMap" or "ArchaMap".}
}
\value{
A list containing detailed information about each dataset, or an error message if the request fails.
The list includes fields such as nodeID, CMName, CMID, shortName, project, Unit, parent, ApplicableYears, and more.
}
\description{
This function retrieves detailed information about all datasets from a specified database.
It returns information such as dataset identifiers, names, applicable years, project details, and other relevant metadata.
}
\examples{
allDatasets(database = "SocioMap")
allDatasets(database = "ArchaMap")
}
