% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{callAPI}
\alias{callAPI}
\title{Call API}
\usage{
callAPI(
  endpoint,
  parameters,
  request = "GET",
  url = "https://catmapper.org/api",
  type = "default"
)
}
\arguments{
\item{endpoint}{API endpoint}

\item{parameters}{API parameters}

\item{request}{GET or POST}

\item{url}{API URL}

\item{type}{default or stream}
}
\value{
API response
}
\description{
This is a helper function to call the API. It takes the endpoint, parameters, request type, and URL as input and returns the API response.
}
\examples{
callAPI(endpoint = "search", parameters = list(term = "Dan", database = "SocioMap", property = "Name", domain = "ETHNICITY"), request = "GET")
}
