% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{datasetInfo}
\alias{datasetInfo}
\title{Retrieve Dataset Information by CMID}
\usage{
datasetInfo(database, CMID, domain = "CATEGORY")
}
\arguments{
\item{database}{A string specifying the database to search in. Valid options are "SocioMap" or "ArchaMap".}

\item{CMID}{The CMID of the dataset to retrieve information for (e.g., "SD1" or "AD1").}

\item{domain}{(Optional) A category to filter dataset relationships. Defaults to "CATEGORY" if not specified.}
}
\value{
A list containing detailed information about the dataset, or an error message if the request fails.
}
\description{
This function retrieves detailed information about a dataset based on a given CMID (CatMapperID) from a specified database, with an optional domain filter.
It fetches relationships and properties of datasets associated with the specified CMID.
}
\examples{
datasetInfo(database = "SocioMap", CMID = "SD1", domain = "CATEGORY")
datasetInfo(database = "ArchaMap", CMID = "AD1")
}
