% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{searchDatabase}
\alias{searchDatabase}
\title{Search}
\usage{
searchDatabase(
  database,
  domain = NULL,
  term = NULL,
  property = "Name",
  yearStart = NULL,
  yearEnd = NULL,
  country = NULL,
  context = NULL,
  limit = 1000
)
}
\arguments{
\item{database}{name of database (SocioMap or ArchaMap)}

\item{domain}{name of category domain to search for (DISTRICT, ETHNICITY, etc.)}

\item{term}{search term ("Afghanistan")}

\item{property}{property to search by (Name, CMID, or Key)}

\item{yearStart}{year to search by start}

\item{yearEnd}{year to search by end}

\item{country}{country to search by (must be a CMID)}

\item{context}{context to search by (e.g., hierarchical category as in the state for a county--must be a CMID)}

\item{limit}{limit number of results}
}
\value{
dataframe
}
\description{
Search for a term in a database and retrieve the CMID (CatMapperID), CMName (CatMapper Name), and other information about the matches.
}
\examples{
searchDatabase(database = "SocioMap",domain = "ETHNICITY", term = "Dan",property = "Name", yearStart = NULL, yearEnd = NULL, context = NULL, limit = 1000)
}
