% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Translate}
\usage{
translate(
  rows,
  database,
  term,
  property = "Name",
  domain = "CATEGORY",
  context = NULL,
  country = NULL,
  dataset = NULL,
  yearStart = NULL,
  yearEnd = NULL,
  key = "false",
  query = "false"
)
}
\arguments{
\item{database}{database to use (SocioMap or ArchaMap)}

\item{term}{column name to translate}

\item{property}{property to search by (Name, CMID, or Key)}

\item{domain}{name of category domain to search for (DISTRICT, ETHNICITY, etc.)}

\item{context}{column name of context to search by (CMID)}

\item{country}{column name of country to search by (CMID)}

\item{yearStart}{year to search by start}

\item{yearEnd}{year to search by end}

\item{query}{return query instead of results}
}
\value{
dataframe
}
\description{
This function translate a dataframe by matching the specified term with the property in the database.
}
\examples{
df = data.frame(country = "Afghanistan")
translate(rows = df, database = "SocioMap",domain = "ADM0", term = "country", property = "Name", yearStart = NULL, yearEnd = NULL, country = NULL, context = NULL, query = 'false')
}
